pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "opencv-python-headless>=4.13.0.92",
    "pandas>=2.3.3",
    "pillow>=12.1.1",
    "pyzbar>=0.1.9",
    "streamlit>=1.55.0",
]
